iahx-topic-query/requirements.txt
Django==4.0.7
mysqlclient==2.1.1
gunicorn==20.1.0
djangorestframework==3.13.1
